+++ services/sources/types.go
@@ -51,7 +51,7 @@ type Service interface {
 	GetSetByQuery(ctx context.Context, query, lg string) (*dodugo.EquipmentSet, error)
 
 	GetAlmanaxByDate(ctx context.Context, date time.Time, language string) (*dodugo.AlmanaxEntry, error)
-	GetAlmanaxByRange(ctx context.Context, daysDuration int32, language string) ([]dodugo.AlmanaxEntry, error)
+	GetAlmanaxByRange(ctx context.Context, daysDuration int64, language string) ([]dodugo.AlmanaxEntry, error)
 
 	ListenGameEvent(handler GameEventHandler)
 }
